requirements.txt
numpy==1.23.5
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
scikit_learn==1.2.0
Pillow==9.2.0
scipy==1.9.1
scikit_image==0.20.0
matplotlib==3.2.2
skimage==0.20.0
